+++ datamodel/openapi/openapi-core-apache/src/main/java/com/sap/cloud/sdk/services/openapi/apache/apiclient/ApiClient.java
@@ -15,59 +15,36 @@ package com.sap.cloud.sdk.services.openapi.apache.apiclient;
 import static com.sap.cloud.sdk.services.openapi.apache.apiclient.DefaultApiResponseHandler.isJsonMime;
 import static lombok.AccessLevel.PRIVATE;
 
-import java.io.File;
-import java.io.IOException;
 import java.net.URLEncoder;
-import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
-import java.nio.charset.UnsupportedCharsetException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
+import java.util.function.UnaryOperator;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
-import org.apache.hc.client5.http.entity.UrlEncodedFormEntity;
-import org.apache.hc.client5.http.entity.mime.MultipartEntityBuilder;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
-import org.apache.hc.client5.http.protocol.HttpClientContext;
-import org.apache.hc.core5.http.ContentType;
-import org.apache.hc.core5.http.HttpEntity;
-import org.apache.hc.core5.http.Method;
-import org.apache.hc.core5.http.NameValuePair;
-import org.apache.hc.core5.http.io.HttpClientResponseHandler;
-import org.apache.hc.core5.http.io.entity.ByteArrayEntity;
-import org.apache.hc.core5.http.io.entity.FileEntity;
-import org.apache.hc.core5.http.io.entity.StringEntity;
-import org.apache.hc.core5.http.io.support.ClassicRequestBuilder;
-import org.apache.hc.core5.http.message.BasicNameValuePair;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
 import com.fasterxml.jackson.annotation.PropertyAccessor;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.google.common.annotations.Beta;
 import com.sap.cloud.sdk.cloudplatform.connectivity.ApacheHttpClient5Accessor;
 import com.sap.cloud.sdk.cloudplatform.connectivity.Destination;
-import com.sap.cloud.sdk.services.openapi.apache.core.OpenApiRequestException;
-import com.sap.cloud.sdk.services.openapi.apache.core.OpenApiResponseListener;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
-import lombok.Getter;
 import lombok.ToString;
 import lombok.With;
+import lombok.experimental.Delegate;
+import lombok.experimental.Tolerate;
 
 /**
  * API client for executing HTTP requests using Apache HttpClient 5.
@@ -75,30 +52,12 @@ import lombok.With;
 @AllArgsConstructor( access = PRIVATE )
 @EqualsAndHashCode
 @ToString
-public class ApiClient
+public class ApiClient implements ApiClientInvoker
 {
-    @Nonnull
-    private final CloseableHttpClient httpClient;
-
     @With
-    @Getter
-    @Nonnull
-    private final String basePath;
+    @Delegate
+    private final ApiClientInvoker invoker;
 
-    @With( onMethod_ = @Beta )
-    @Nonnull
-    private final ObjectMapper objectMapper;
-
-    @With
-    @Nullable
-    private final String tempFolderPath;
-
-    @With( onMethod_ = @Beta )
-    @Nullable
-    private final OpenApiResponseListener openApiResponseListener;
-
-    // Methods that can have a request body
-    private static final Set<Method> BODY_METHODS = Set.of(Method.POST, Method.PUT, Method.PATCH, Method.DELETE);
     private static final String DEFAULT_BASE_PATH = "http://localhost";
 
     /**
@@ -111,7 +70,36 @@ public class ApiClient
     @Nonnull
     public static ApiClient fromHttpClient( @Nonnull final CloseableHttpClient httpClient )
     {
-        return new ApiClient(httpClient, DEFAULT_BASE_PATH, createDefaultObjectMapper(), null, null);
+        return fromHttpClient(httpClient, DEFAULT_BASE_PATH);
+    }
+
+    /**
+     * Creates an ApiClient instance from an existing HttpClient.
+     *
+     * @param httpClient
+     *            The HttpClient to use for requests
+     * @param basePath
+     *            The base path to use for requests
+     * @return A new ApiClient instance
+     */
+    @Nonnull
+    public static ApiClient fromHttpClient( @Nonnull final CloseableHttpClient httpClient, @Nonnull String basePath )
+    {
+        return new ApiClient(
+            new ApiClientInvokerDefault(httpClient, basePath, createDefaultObjectMapper(), null, null));
+    }
+
+    /**
+     * Returns a new ApiClient instance with the invoker modified by the given operator.
+     *
+     * @param invokerOperator
+     *            The operator to modify the invoker
+     * @return A new ApiClient instance with the modified invoker
+     */
+    @Tolerate
+    public ApiClient withInvoker( @Nonnull final UnaryOperator<ApiClientInvoker> invokerOperator )
+    {
+        return new ApiClient(invokerOperator.apply(this.invoker));
     }
 
     /**
@@ -124,9 +112,9 @@ public class ApiClient
     @Nonnull
     public static ApiClient create( @Nonnull final Destination destination )
     {
-        return fromHttpClient((CloseableHttpClient) ApacheHttpClient5Accessor.getHttpClient(destination))
-            .withBasePath(destination.asHttp().getUri().toString());
-
+        final CloseableHttpClient httpClient =
+            (CloseableHttpClient) ApacheHttpClient5Accessor.getHttpClient(destination);
+        return fromHttpClient(httpClient, destination.asHttp().getUri().toString());
     }
 
     /**
@@ -331,251 +319,5 @@ public class ApiClient
     public static String escapeString( @Nonnull final String str )
     {
         return URLEncoder.encode(str, StandardCharsets.UTF_8).replaceAll("\\+", "%20");
-    }
-
-    /**
-     * Parse content type object from header value
-     */
-    @Nonnull
-    private ContentType getContentType( @Nonnull final String headerValue )
-        throws OpenApiRequestException
-    {
-        try {
-            return ContentType.parse(headerValue);
-        }
-        catch( UnsupportedCharsetException e ) {
-            throw new OpenApiRequestException("Could not parse content type " + headerValue, e);
-        }
-    }
-
-    /**
-     * Serialize the given Java object into string according the given Content-Type (only JSON is supported for now).
-     *
-     * @param obj
-     *            Object
-     * @param contentType
-     *            Content type
-     * @param formParams
-     *            Form parameters
-     * @return Object
-     * @throws OpenApiRequestException
-     *             API exception
-     */
-    @Nonnull
-    private HttpEntity serialize(
-        @Nullable final Object obj,
-        @Nonnull final Map<String, Object> formParams,
-        @Nonnull final ContentType contentType )
-        throws OpenApiRequestException
-    {
-        final String mimeType = contentType.getMimeType();
-        if( isJsonMime(mimeType) ) {
-            try {
-                return new StringEntity(
-                    objectMapper.writeValueAsString(obj),
-                    contentType.withCharset(StandardCharsets.UTF_8));
-            }
-            catch( JsonProcessingException e ) {
-                throw new OpenApiRequestException(e);
-            }
-        } else if( mimeType.equals(ContentType.MULTIPART_FORM_DATA.getMimeType()) ) {
-            final MultipartEntityBuilder multiPartBuilder = MultipartEntityBuilder.create();
-            for( final Entry<String, Object> paramEntry : formParams.entrySet() ) {
-                final Object value = paramEntry.getValue();
-                if( value instanceof File file ) {
-                    multiPartBuilder.addBinaryBody(paramEntry.getKey(), file);
-                } else if( value instanceof byte[] byteArray ) {
-                    multiPartBuilder.addBinaryBody(paramEntry.getKey(), byteArray);
-                } else {
-                    final Charset charset = contentType.getCharset();
-                    if( charset != null ) {
-                        final ContentType customContentType =
-                            ContentType.create(ContentType.TEXT_PLAIN.getMimeType(), charset);
-                        multiPartBuilder
-                            .addTextBody(
-                                paramEntry.getKey(),
-                                parameterToString(paramEntry.getValue()),
-                                customContentType);
-                    } else {
-                        multiPartBuilder.addTextBody(paramEntry.getKey(), parameterToString(paramEntry.getValue()));
-                    }
-                }
-            }
-            return multiPartBuilder.build();
-        } else if( mimeType.equals(ContentType.APPLICATION_FORM_URLENCODED.getMimeType()) ) {
-            final List<NameValuePair> formValues = new ArrayList<>();
-            for( final Entry<String, Object> paramEntry : formParams.entrySet() ) {
-                formValues.add(new BasicNameValuePair(paramEntry.getKey(), parameterToString(paramEntry.getValue())));
-            }
-            return new UrlEncodedFormEntity(formValues, contentType.getCharset());
-        } else {
-            // Handle files with unknown content type
-            if( obj instanceof File file ) {
-                return new FileEntity(file, contentType);
-            } else if( obj instanceof byte[] byteArray ) {
-                return new ByteArrayEntity(byteArray, contentType);
-            }
-            throw new OpenApiRequestException("Serialization for content type '" + contentType + "' not supported");
-        }
-    }
-
-    /**
-     * Build full URL by concatenating base URL, the given sub path and query parameters.
-     *
-     * @param path
-     *            The sub path
-     * @param queryParams
-     *            The query parameters
-     * @param collectionQueryParams
-     *            The collection query parameters
-     * @param urlQueryDeepObject
-     *            URL query string of the deep object parameters
-     * @return The full URL
-     */
-    @Nonnull
-    private String buildUrl(
-        @Nonnull final String path,
-        @Nullable final List<Pair> queryParams,
-        @Nullable final List<Pair> collectionQueryParams,
-        @Nullable final String urlQueryDeepObject )
-    {
-        final StringBuilder url = new StringBuilder();
-        if( basePath.endsWith("/") && path.startsWith("/") ) {
-            url.append(basePath, 0, basePath.length() - 1);
-        } else {
-            url.append(basePath);
-        }
-        url.append(path);
-
-        if( queryParams != null && !queryParams.isEmpty() ) {
-            // support (constant) query string in `path`, e.g. "/posts?draft=1"
-            String prefix = path.contains("?") ? "&" : "?";
-            for( final Pair param : queryParams ) {
-                if( prefix != null ) {
-                    url.append(prefix);
-                    prefix = null;
-                } else {
-                    url.append("&");
-                }
-                final String value = parameterToString(param.getValue());
-                // query parameter value already escaped as part of parameterToPair
-                url.append(escapeString(param.getName())).append("=").append(value);
-            }
-        }
-
-        if( collectionQueryParams != null && !collectionQueryParams.isEmpty() ) {
-            String prefix = url.toString().contains("?") ? "&" : "?";
-            for( final Pair param : collectionQueryParams ) {
-                if( prefix != null ) {
-                    url.append(prefix);
-                    prefix = null;
-                } else {
-                    url.append("&");
-                }
-                final String value = parameterToString(param.getValue());
-                // collection query parameter value already escaped as part of parameterToPairs
-                url.append(escapeString(param.getName())).append("=").append(value);
-            }
-        }
-
-        if( urlQueryDeepObject != null && !urlQueryDeepObject.isEmpty() ) {
-            url.append(url.toString().contains("?") ? "&" : "?");
-            url.append(urlQueryDeepObject);
-        }
-
-        return url.toString();
-    }
-
-    private static boolean isBodyAllowed( @Nonnull final Method method )
-    {
-        return BODY_METHODS.contains(method);
-    }
-
-    /**
-     * Invoke API by sending HTTP request with the given options.
-     *
-     * @param <T>
-     *            Type
-     * @param path
-     *            The sub-path of the HTTP URL
-     * @param method
-     *            The request method, one of "GET", "POST", "PUT", and "DELETE"
-     * @param queryParams
-     *            The query parameters
-     * @param collectionQueryParams
-     *            The collection query parameters
-     * @param urlQueryDeepObject
-     *            A URL query string for deep object parameters
-     * @param body
-     *            The request body object - if it is not binary, otherwise null
-     * @param headerParams
-     *            The header parameters
-     * @param formParams
-     *            The form parameters
-     * @param accept
-     *            The request's Accept header
-     * @param contentType
-     *            The request's Content-Type header
-     * @param returnType
-     *            Return type
-     * @return The response body in type of string
-     * @throws OpenApiRequestException
-     *             API exception
-     */
-    @Beta
-    @Nullable
-    public <T> T invokeAPI(
-        @Nonnull final String path,
-        @Nonnull final String method,
-        @Nullable final List<Pair> queryParams,
-        @Nullable final List<Pair> collectionQueryParams,
-        @Nullable final String urlQueryDeepObject,
-        @Nullable final Object body,
-        @Nonnull final Map<String, String> headerParams,
-        @Nonnull final Map<String, Object> formParams,
-        @Nullable final String accept,
-        @Nonnull final String contentType,
-        @Nonnull final TypeReference<T> returnType )
-        throws OpenApiRequestException
-    {
-        if( body != null && !formParams.isEmpty() ) {
-            throw new OpenApiRequestException("Cannot have body and form params");
-        }
-
-        final String url = buildUrl(path, queryParams, collectionQueryParams, urlQueryDeepObject);
-
-        final ClassicRequestBuilder builder = ClassicRequestBuilder.create(method);
-        builder.setUri(url);
-
-        if( accept != null ) {
-            builder.addHeader("Accept", accept);
-        }
-        for( final Entry<String, String> keyValue : headerParams.entrySet() ) {
-            builder.addHeader(keyValue.getKey(), keyValue.getValue());
-        }
-
-        final HttpClientContext context = HttpClientContext.create();
-
-        final ContentType contentTypeObj = getContentType(contentType);
-        if( body != null || !formParams.isEmpty() ) {
-            if( isBodyAllowed(Method.valueOf(method)) ) {
-                // Add entity if we have content and a valid method
-                builder.setEntity(serialize(body, formParams, contentTypeObj));
-            } else {
-                throw new OpenApiRequestException("method " + method + " does not support a request body");
-            }
-        } else {
-            // for empty body
-            builder.setEntity(new StringEntity("", contentTypeObj));
-        }
-
-        try {
-            final HttpClientResponseHandler<T> responseHandler =
-                new DefaultApiResponseHandler<>(objectMapper, tempFolderPath, returnType, openApiResponseListener);
-            return httpClient.execute(builder.build(), context, responseHandler);
-        }
-        catch( IOException e ) {
-            throw new OpenApiRequestException(e);
-        }
     }
 }
